feat: add interaction logic and UI templates for Quick Insight and Deep Dive analysis creation flows
